requirements.txt
pytest==5.4.*
selenium==4.0.0a6.post2
envparse==0.2.*
allure-pytest==2.8.*
